Move FindCommand and NameContainsKeywordsPredicate
