+++ src/main/java/seedu/address/model/student/predicates/NameContainsKeywordsPredicate.java
@@ -1,9 +1,10 @@
-package seedu.address.model.student;
+package seedu.address.model.student.predicates;
 
 import java.util.List;
 import java.util.function.Predicate;
 
 import seedu.address.commons.util.StringUtil;
+import seedu.address.model.student.Student;
 
 /**
  * Tests that a {@code Person}'s {@code Name} matches any of the keywords given.
